danguard-v3.1_u/requirements.txt
python-dotenv
bcrypt
requests
dnspython
azure-identity
azure-mgmt-web
boto3
google-api-python-client
google-auth-oauthlib